Revert "Linearalgebra"
